requirements.txt
# Core dependencies
python-binance==1.0.19
torch>=2.0.0
pandas>=2.0.0
numpy>=1.24.0
pandas-ta>=0.3.14b0
python-dotenv>=1.0.0
python-telegram-bot>=20.0
requests>=2.31.0
scikit-learn>=1.3.0

# Optional: TA-Lib (requires manual installation on Windows)
# ta-lib==0.4.28

